Squidqueen Bounty - Nosferatu Hair and Ear, Gargoyle Hair changes (#752)
## About The Pull Request

This was done as a bounty from Dolly - Whispers - Asim - Aegis
(squidqueen) on the discord since they wanted to give Gargoyles and
Nosferatu all of the hair styles as well as allowing you to choose to
have pointy ears or not on Nosferatu. Regardless of choice they'll still
have fucked up skin and huge fangs. If someone wants to make comments or
suggestions about it, hit them up in the bounty thread.

NOTE: There is already a PR for the Nosferatu hair stuff here
https://github.com/WorldOfDarknessXIII/World-of-Darkness-13/pull/750
The changes done to the Nosferatu file in this PR are the exact same
with the exception of the ear options. Keep the differences in mind
if/when merging either of them.

## Why It's Good For The Game

They wanted more customization options which is always nice.

## Changelog

:cl:
add: All hair styles for Gargoyles and Nosferatu can now be selected. Do
try to keep both in character though with grey colored hair that could
reasonably be on the statues and wretched lookin ones for the nossies.
image: Modified the Nosferatu head sprites so their ears aren't
permanently on. Honestly this probably should have been done anyways
considering that their ears are already a separate icon.
/:cl: